g502: removed workqueue usage and small improvements.
The driver doesn't work yet, though it's about to be finished until I
find the main problem that causes BTN6 not to work, as well as not
receiving parseable data in g502_handle_regular_event.

I barely have any time to work on the module/driver due to the military.

Signed-off-by: Roi L <roiinpushq@proton.me>