g502: overall refactor, improved performance and readability
- fixed a bug where DPI_SET feature wasn't working, due to unvalid
  feature function index.
- do not use spinlocks to time hid_raw requests, mutex instead.
- report_fixup has won the const qualifier since 6.12 so I have made
it compatiable for users under 6.12.
- Remove the
sysfs group before the module is unloaded
and some small
improvements to the efficienty of the program

Instead of using a virtual machine with a virtualized usbhub, I am now testing my driver on a real machine, my laptop,  and it
clearly has several more benefits to debugging and to the driver
itself.

Signed-off-by: Roi L <roiinpushq@proton.me>